Move eservice-template-process tests to integration dir
