+++ packages/eservice-template-process/test/integration/createEServiceTemplateRiskAnalysis.test.ts
@@ -40,14 +40,14 @@ import {
   templateNotInReceiveMode,
   tenantKindNotFound,
   tenantNotFound,
-} from "../src/model/domain/errors.js";
+} from "../../src/model/domain/errors.js";
 import {
   addOneEServiceTemplate,
-  buildRiskAnalysisSeed,
   eserviceTemplateService,
   readLastEserviceTemplateEvent,
   addOneTenant,
-} from "./utils.js";
+} from "../integrationUtils.js";
+import { buildRiskAnalysisSeed } from "../mockUtils.js";
 
 describe("createEServiceTemplateRiskAnalysis", () => {
   const mockDate = new Date();
