+++ packages/eservice-template-process/test/integration/getEServiceTemplateCreators.test.ts
@@ -16,7 +16,7 @@ import {
   addOneEServiceTemplate,
   addOneTenant,
   eserviceTemplateService,
-} from "./utils.js";
+} from "../integrationUtils.js";
 
 describe("getEServiceTemplateCreators", () => {
   let tenant1: Tenant;
