+++ packages/eservice-template-process/test/integration/updateEServiceTemplateIntendedTarget.test.ts
@@ -9,28 +9,25 @@ import {
 } from "pagopa-interop-commons-test";
 import {
   operationForbidden,
-  generateId,
-  TenantId,
   EServiceTemplateVersion,
   eserviceTemplateVersionState,
   EServiceTemplate,
   toEServiceTemplateV2,
-  EServiceTemplateNameUpdatedV2,
+  EServiceTemplateIntendedTargetUpdatedV2,
 } from "pagopa-interop-models";
 import { expect, describe, it } from "vitest";
 import {
   eserviceTemplateWithoutPublishedVersion,
   eServiceTemplateNotFound,
-  eServiceTemplateDuplicate,
-} from "../src/model/domain/errors.js";
+} from "../../src/model/domain/errors.js";
 import {
   addOneEServiceTemplate,
   eserviceTemplateService,
   readLastEserviceTemplateEvent,
-} from "./utils.js";
+} from "../integrationUtils.js";
 
-describe("updateEServiceTemplateName", () => {
-  it("should write on event-store for the update of the eService template name", async () => {
+describe("updateEServiceTemplateIntendedTarget", () => {
+  it("should write on event-store for the update of the eService template description", async () => {
     const eserviceTemplateVersion: EServiceTemplateVersion = {
       ...getMockEServiceTemplateVersion(),
       state: eserviceTemplateVersionState.published,
@@ -41,18 +38,18 @@ describe("updateEServiceTemplateName", () => {
       versions: [eserviceTemplateVersion],
     };
     await addOneEServiceTemplate(eserviceTemplate);
-    const updatedName = "eservice template new name";
+    const updatedTemplateDescription = "eservice template new description";
     const returnedEServiceTemplate =
-      await eserviceTemplateService.updateEServiceTemplateName(
+      await eserviceTemplateService.updateEServiceTemplateIntendedTarget(
         eserviceTemplate.id,
-        updatedName,
+        updatedTemplateDescription,
         getMockContext({
           authData: getMockAuthData(eserviceTemplate.creatorId),
         })
       );
     const updatedEServiceTemplate: EServiceTemplate = {
       ...eserviceTemplate,
-      name: updatedName,
+      intendedTarget: updatedTemplateDescription,
     };
     const writtenEvent = await readLastEserviceTemplateEvent(
       eserviceTemplate.id
@@ -60,11 +57,11 @@ describe("updateEServiceTemplateName", () => {
     expect(writtenEvent).toMatchObject({
       stream_id: eserviceTemplate.id,
       version: "1",
-      type: "EServiceTemplateNameUpdated",
+      type: "EServiceTemplateIntendedTargetUpdated",
       event_version: 2,
     });
     const writtenPayload = decodeProtobufPayload({
-      messageType: EServiceTemplateNameUpdatedV2,
+      messageType: EServiceTemplateIntendedTargetUpdatedV2,
       payload: writtenEvent.data,
     });
     expect(writtenPayload.eserviceTemplate).toEqual(
@@ -78,9 +75,9 @@ describe("updateEServiceTemplateName", () => {
   it("should throw eServiceTemplateNotFound if the eservice template doesn't exist", async () => {
     const eserviceTemplate = getMockEServiceTemplate();
     expect(
-      eserviceTemplateService.updateEServiceTemplateName(
+      eserviceTemplateService.updateEServiceTemplateIntendedTarget(
         eserviceTemplate.id,
-        "eservice template new name",
+        "eservice template new description",
         getMockContext({
           authData: getMockAuthData(eserviceTemplate.creatorId),
         })
@@ -91,9 +88,9 @@ describe("updateEServiceTemplateName", () => {
     const eserviceTemplate = getMockEServiceTemplate();
     await addOneEServiceTemplate(eserviceTemplate);
     expect(
-      eserviceTemplateService.updateEServiceTemplateName(
+      eserviceTemplateService.updateEServiceTemplateIntendedTarget(
         eserviceTemplate.id,
-        "eservice template new name",
+        "eservice template new description",
         getMockContext({})
       )
     ).rejects.toThrowError(operationForbidden);
@@ -102,9 +99,9 @@ describe("updateEServiceTemplateName", () => {
     const eserviceTemplate = getMockEServiceTemplate();
     await addOneEServiceTemplate(eserviceTemplate);
     expect(
-      eserviceTemplateService.updateEServiceTemplateName(
+      eserviceTemplateService.updateEServiceTemplateIntendedTarget(
         eserviceTemplate.id,
-        "eservice template new name",
+        "eservice template new description",
         getMockContext({
           authData: getMockAuthData(eserviceTemplate.creatorId),
         })
@@ -125,9 +122,9 @@ describe("updateEServiceTemplateName", () => {
     };
     await addOneEServiceTemplate(eserviceTemplate);
     expect(
-      eserviceTemplateService.updateEServiceTemplateName(
+      eserviceTemplateService.updateEServiceTemplateIntendedTarget(
         eserviceTemplate.id,
-        "eservice template new name",
+        "eservice template new description",
         getMockContext({
           authData: getMockAuthData(eserviceTemplate.creatorId),
         })
@@ -135,41 +132,5 @@ describe("updateEServiceTemplateName", () => {
     ).rejects.toThrowError(
       eserviceTemplateWithoutPublishedVersion(eserviceTemplate.id)
     );
-  });
-  it("should throw eServiceTemplateDuplicate is there is another eservice template with the same name by the same creator", async () => {
-    const creatorId = generateId<TenantId>();
-
-    const eserviceTemplateVersion: EServiceTemplateVersion = {
-      ...getMockEServiceTemplateVersion(),
-      interface: getMockDocument(),
-      state: eserviceTemplateVersionState.published,
-    };
-    const eserviceTemplate: EServiceTemplate = {
-      ...getMockEServiceTemplate(),
-      creatorId,
-      versions: [eserviceTemplateVersion],
-    };
-
-    const duplicateName = "eservice duplciate name";
-
-    const eserviceTemplateWithSameName: EServiceTemplate = {
-      ...getMockEServiceTemplate(),
-      creatorId,
-      name: duplicateName,
-    };
-
-    await addOneEServiceTemplate(eserviceTemplate);
-    await addOneEServiceTemplate(eserviceTemplateWithSameName);
-
-    const updatedName = duplicateName;
-    expect(
-      eserviceTemplateService.updateEServiceTemplateName(
-        eserviceTemplate.id,
-        updatedName,
-        getMockContext({
-          authData: getMockAuthData(eserviceTemplate.creatorId),
-        })
-      )
-    ).rejects.toThrowError(eServiceTemplateDuplicate(duplicateName));
   });
 });
