+++ packages/eservice-template-process/test/integration/updateEServiceTemplateVersionQuotas.test.ts
@@ -15,40 +15,42 @@ import {
   EServiceTemplate,
   EServiceTemplateVersion,
   eserviceTemplateVersionState,
-  EServiceTemplateVersionActivatedV2,
+  EServiceTemplateVersionQuotasUpdatedV2,
 } from "pagopa-interop-models";
 import { expect, describe, it } from "vitest";
 import {
   eServiceTemplateNotFound,
   eServiceTemplateVersionNotFound,
+  inconsistentDailyCalls,
   notValidEServiceTemplateVersionState,
-} from "../src/model/domain/errors.js";
+} from "../../src/model/domain/errors.js";
 import {
   eserviceTemplateService,
   addOneEServiceTemplate,
   readLastEserviceTemplateEvent,
-} from "./utils.js";
+} from "../integrationUtils.js";
 
-describe("activateEServiceTemplateVersion", () => {
+describe("updateEServiceTemplateVersionQuotas", () => {
   const mockEServiceTemplate = getMockEServiceTemplate();
   const mockEServiceTemplateVersion = getMockEServiceTemplateVersion();
   const mockDocument = getMockDocument();
 
-  it("should write on event-store for the activation of a eservice template version", async () => {
+  it("should write on event-store for the update of the eservice template version quotas", async () => {
     const eserviceTemplateVersion: EServiceTemplateVersion = {
       ...mockEServiceTemplateVersion,
       interface: mockDocument,
-      state: descriptorState.suspended,
-      suspendedAt: new Date(),
+      state: descriptorState.published,
     };
     const eserviceTemplate: EServiceTemplate = {
       ...mockEServiceTemplate,
       versions: [eserviceTemplateVersion],
     };
     await addOneEServiceTemplate(eserviceTemplate);
-    await eserviceTemplateService.activateEServiceTemplateVersion(
+
+    await eserviceTemplateService.updateEServiceTemplateVersionQuotas(
       eserviceTemplate.id,
       eserviceTemplateVersion.id,
+      { voucherLifespan: 60 },
       getMockContext({
         authData: getMockAuthData(eserviceTemplate.creatorId),
       })
@@ -59,10 +61,10 @@ describe("activateEServiceTemplateVersion", () => {
     );
     expect(writtenEvent.stream_id).toBe(eserviceTemplate.id);
     expect(writtenEvent.version).toBe("1");
-    expect(writtenEvent.type).toBe("EServiceTemplateVersionActivated");
+    expect(writtenEvent.type).toBe("EServiceTemplateVersionQuotasUpdated");
     expect(writtenEvent.event_version).toBe(2);
     const writtenPayload = decodeProtobufPayload({
-      messageType: EServiceTemplateVersionActivatedV2,
+      messageType: EServiceTemplateVersionQuotasUpdatedV2,
       payload: writtenEvent.data,
     });
 
@@ -71,8 +73,7 @@ describe("activateEServiceTemplateVersion", () => {
       versions: [
         {
           ...eserviceTemplateVersion,
-          state: descriptorState.published,
-          suspendedAt: undefined,
+          voucherLifespan: 60,
         },
       ],
     });
@@ -85,9 +86,10 @@ describe("activateEServiceTemplateVersion", () => {
 
   it("should throw eServiceTemplateNotFound if the eservice template doesn't exist", () => {
     expect(
-      eserviceTemplateService.activateEServiceTemplateVersion(
+      eserviceTemplateService.updateEServiceTemplateVersionQuotas(
         mockEServiceTemplate.id,
         mockEServiceTemplateVersion.id,
+        { voucherLifespan: 60 },
         getMockContext({
           authData: getMockAuthData(mockEServiceTemplate.creatorId),
         })
@@ -107,9 +109,10 @@ describe("activateEServiceTemplateVersion", () => {
     };
     await addOneEServiceTemplate(eserviceTemplate);
     expect(
-      eserviceTemplateService.activateEServiceTemplateVersion(
+      eserviceTemplateService.updateEServiceTemplateVersionQuotas(
         eserviceTemplate.id,
         eserviceTemplateVersion.id,
+        { voucherLifespan: 60 },
         getMockContext({})
       )
     ).rejects.toThrowError(operationForbidden);
@@ -123,11 +126,12 @@ describe("activateEServiceTemplateVersion", () => {
     await addOneEServiceTemplate(eserviceTemplate);
 
     expect(
-      eserviceTemplateService.activateEServiceTemplateVersion(
+      eserviceTemplateService.updateEServiceTemplateVersionQuotas(
         eserviceTemplate.id,
         mockEServiceTemplateVersion.id,
+        { voucherLifespan: 60 },
         getMockContext({
-          authData: getMockAuthData(eserviceTemplate.creatorId),
+          authData: getMockAuthData(mockEServiceTemplate.creatorId),
         })
       )
     ).rejects.toThrowError(
@@ -140,7 +144,6 @@ describe("activateEServiceTemplateVersion", () => {
 
   it.each([
     eserviceTemplateVersionState.draft,
-    descriptorState.published,
     eserviceTemplateVersionState.deprecated,
   ])(
     "should throw notValidEServiceTemplateVersionState if the descriptor is in %s state",
@@ -155,9 +158,10 @@ describe("activateEServiceTemplateVersion", () => {
       };
       await addOneEServiceTemplate(eserviceTemplate);
       expect(
-        eserviceTemplateService.activateEServiceTemplateVersion(
+        eserviceTemplateService.updateEServiceTemplateVersionQuotas(
           eserviceTemplate.id,
           eserviceTemplateVersion.id,
+          { voucherLifespan: 60 },
           getMockContext({
             authData: getMockAuthData(eserviceTemplate.creatorId),
           })
@@ -167,4 +171,32 @@ describe("activateEServiceTemplateVersion", () => {
       );
     }
   );
+
+  it("should throw inconsistentDailyCalls if dailyCallsPerConsumer is greater than dailyCallsTotal", async () => {
+    const eserviceTemplateVersion: EServiceTemplateVersion = {
+      ...mockEServiceTemplateVersion,
+      interface: mockDocument,
+      state: descriptorState.published,
+    };
+    const eserviceTemplate: EServiceTemplate = {
+      ...mockEServiceTemplate,
+      versions: [eserviceTemplateVersion],
+    };
+    await addOneEServiceTemplate(eserviceTemplate);
+
+    expect(
+      eserviceTemplateService.updateEServiceTemplateVersionQuotas(
+        eserviceTemplate.id,
+        mockEServiceTemplateVersion.id,
+        {
+          voucherLifespan: 60,
+          dailyCallsPerConsumer: 11,
+          dailyCallsTotal: 10,
+        },
+        getMockContext({
+          authData: getMockAuthData(mockEServiceTemplate.creatorId),
+        })
+      )
+    ).rejects.toThrowError(inconsistentDailyCalls());
+  });
 });
